% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jayhawkdown.R
\docType{package}
\name{jayhawkdown}
\alias{jayhawkdown}
\title{jayhawkdown: Theses and dissertations for the University of Kansas}
\description{
This package creates theses/dissertations for the University of Kansas.
}
\section{thesis_gitbook}{

Creates an R Markdown thesis template as a webpage
}

\section{thesis_pdf}{

Creates an R Markdown thesis template as a PDF
}

