% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{thesis_pdf}
\alias{thesis_pdf}
\title{Creates an R Markdown PDF Thesis Document}
\usage{
thesis_pdf(toc = TRUE, toc_depth = 3, ...)
}
\arguments{
\item{toc}{A Boolean (TRUE or FALSE) specifying where table of contents
should be created}

\item{toc_depth}{A positive integer}

\item{...}{arguments to be passed to \code{\link[bookdown:pdf_book]{bookdown::pdf_book()}}}
}
\value{
A modified \code{pdf_document} based on the Reed Senior Thesis LaTeX
template
}
\description{
This is a function called in output in the YAML of the driver Rmd file
to specify using the University of Kansas LaTeX template and cls files.
}
\note{
The arguments highlight, keep_tex, and pandoc_args, are already set.
}
\examples{
\dontrun{
 output: jayhawkdown::thesis_pdf
}
}
